requirements.txt
fastmcp
flask
openai
python-dotenv
chromadb
